chore: move enum.js to lib folder (#2285)
