+++ src/bundles/ipfs-provider.js
@@ -3,7 +3,7 @@ import { multiaddr } from '@multiformats/multiaddr'
 import { getIpfs, providers } from 'ipfs-provider'
 import first from 'it-first'
 import last from 'it-last'
-import * as Enum from './enum.js'
+import * as Enum from '../lib/enum.js'
 import { perform } from './task.js'
 import { readSetting, writeSetting } from './local-storage.js'
