+++ src/lib/enum.js
